agregue alertas

+++ js/carrito.js
@@ -1,9 +1,14 @@
+
 const productosEnCarrito = JSON.parse(localStorage.getItem("productos-en-carrito"));
 
 document.addEventListener("DOMContentLoaded", () => {
   const contenedorProductos = document.querySelector("#carrito-Viajes");
 
   function cargarProductosCarrito() {
+
+
+
+
     if (productosEnCarrito && productosEnCarrito.length > 0) {
       productosEnCarrito.forEach((producto, i) => {
         const div = document.createElement("div");
@@ -11,14 +16,14 @@ document.addEventListener("DOMContentLoaded", () => {
         div.innerHTML = `
         <div class="cont-img-carrito">
         <h3>Usted ha reservado un vuelo:</h3>
-        <p>Origen: <span class="compra-origen">${producto.origen}</span></p>
-        <p>Destino: <span class="compra-destino">${producto.destino}</span></p>
+        <p class="compra-origen"> Origen: <span class="compra-origen">${producto.origen}</span></p>
+        <p class="compra-origen" >Destino: <span class="compra-destino">${producto.destino}</span></p>
         <p>Descripcion: ${producto.descripcion} </p>
-        <p>Fecha de partida: <span class="compra-fecha-partida">20/05/2018</span> Hora: <span
+        <p>Fecha de partida: <span class="compra-fecha-partida">12/06/2018</span> Hora: <span
               class="compra-hora-partida">10:00</span></p>
         <p>Fecha de regreso: <span class="compra-fecha-regreso">22/05/2018</span> Hora: <span
               class="compra-hora-regreso">16:00</span></p>
-              <p class="compra-personas">Cantidad de personas: <input type="number" class="compra-input-personas"
+              <p class="compra-personas">Cantidad de personas: <input id="miInput" type="number" class="compra-input-personas"
               min="1" max="10" value="${producto.cantidadPersonas}"></p>
           <!-- ... -->
         <p class="compra-equipaje">¿Llevará equipaje? <select class="compra-input-equipaje">
@@ -49,26 +54,65 @@ document.addEventListener("DOMContentLoaded", () => {
             <option value="no" selected>No</option>
           </select>
         </p>
-       
+        <h5>Precio del viaje Por Persona: <span class="compra-precio">${producto.precio}</span></h5>
         <img class="imgcarritook" src="${producto.imagen2}" alt="destino">
        
 
-        <h5>Precio del viaje Por Persona: <span class="compra-precio">${producto.precio}</span></h5>
-        <h5>Precio Total del viaje: <span class="compra-precio-total">${producto.precio * producto.precio}</span></h5>
+        
+        <h5 class ="totviaje">Precio Total del viaje: <span class="compra-precio-total">       </span></h5>
           <button class="eliminar-vuelo" data-index="${i}">Eliminar Vuelo</button>
-          <button class="descargar-boletos">Descargar Boletos</button>
+         
           </div>
         `;
         contenedorProductos.appendChild(div);
 
         const botonEliminar = div.querySelector(".eliminar-vuelo");
         botonEliminar.addEventListener("click", () => {
-          const index = botonEliminar.dataset.index;
-          productosEnCarrito.splice(index, 1);
-          localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito));
-          div.remove();
+
+
+          Swal.fire({
+            title: '¿Seguro que quieres eliminar esta reserva?',
+            text: "Si lo borras tendrás que agregarlo nuevamente.",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Si, Borrar!'
+          }).then((result) => {
+
+            if (result.isConfirmed) {
+
+              const index = botonEliminar.dataset.index;
+              productosEnCarrito.splice(index, 1);
+              localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito));
+              div.remove();
+    
+              Swal.fire(
+                'Eliminado!',
+                'La reserva ha sido eliminada correctamente.',
+                'success'
+              )
+            }
+          })
         });
       });
+
+      const miInput = document.getElementById("miInput");
+
+      miInput.addEventListener("change", () => {
+        const cantidad = miInput.value;
+        const div = miInput.parentNode.parentNode; // Obtener el div padre
+        const precioSpan = div.querySelector(".compra-precio");
+        const precioTotalSpan = div.querySelector(".compra-precio-total");
+        const precio = parseFloat(precioSpan.textContent); // Obtener el precio como número
+        const precioTotal = precio * cantidad;
+        precioTotalSpan.textContent = precioTotal;
+        const elPrecioTotal = document.getElementById("elPrecioTotal")
+        elPrecioTotal.textContent = precioTotal;
+      });
+
+
+
 
       actualizarBotonesEliminar();
       actualizarTotal();
@@ -81,7 +125,7 @@ document.addEventListener("DOMContentLoaded", () => {
 
 
 
-
+//alquiler de autos..................................................................................
 
 
 document.addEventListener("DOMContentLoaded", () => {
@@ -91,6 +135,7 @@ document.addEventListener("DOMContentLoaded", () => {
     const autosEnCarrito = JSON.parse(localStorage.getItem("autos-en-carrito"));
 
     if (autosEnCarrito && autosEnCarrito.length > 0) {
+
       autosEnCarrito.forEach((auto, i) => {
         const div = document.createElement("div");
         div.classList.add("izquierdo");
@@ -166,10 +211,33 @@ document.addEventListener("DOMContentLoaded", () => {
 
         const botonEliminar = div.querySelector(".eliminar-vuelo");
         botonEliminar.addEventListener("click", () => {
-          const index = botonEliminar.dataset.index;
-          productosEnCarrito.splice(index, 1);
-          localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito));
-          div.remove();
+
+          Swal.fire({
+            title: 'Seguro Que Quieres Eliminar Este Vehículo',
+            text: "¿Estás seguro de que deseas eliminar esta reserva? Si la eliminas, tendrás que hacer una nueva reserva para alquilar este vehículo. ",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Si, Borrar!'
+          }).then((result) => {
+            if (result.isConfirmed) {
+
+              const index = botonEliminar.dataset.index;
+              autosEnCarrito.splice(index, 1);
+              localStorage.setItem("autos-en-carrito", JSON.stringify(autosEnCarrito));
+              div.remove();
+
+              Swal.fire(
+                'Reserva Eliminada!',
+                'Tu reserva ha sido eliminada correctamente.',
+                'success'
+              )
+            }
+          })
+
+
+         
         });
 
       });
